requirements.txt
pandas
PyMySQL
mysql-replication
mysql-connector-python